refactor: split doctor checks
