+++ src/daemon/handlers/session-doctor-options.ts
@@ -1,0 +1,120 @@
+import { detectProjectRuntimeKind } from '../../utils/project-runtime.ts';
+import type { SessionStore } from '../session-store.ts';
+import type { DaemonRequest, SessionState } from '../types.ts';
+import type { DoctorCheck, DoctorKind, DoctorOptions } from './session-doctor-types.ts';
+
+const DEFAULT_METRO_HOST = '127.0.0.1';
+const DEFAULT_METRO_PORT = 8081;
+const REMOTE_CONNECTION_FLAG_KEYS = ['daemonBaseUrl', 'tenant', 'runId', 'leaseId'] as const;
+
+export function readDoctorOptions(
+  req: DaemonRequest,
+  session: SessionState | undefined,
+): DoctorOptions {
+  const kind = detectProjectRuntimeKind(req.meta?.cwd);
+  const targetApp = session?.appBundleId;
+  const metroHost = readNonEmptyString(req.runtime?.metroHost) ?? DEFAULT_METRO_HOST;
+  const metroPort = readPositivePort(req.runtime?.metroPort) ?? DEFAULT_METRO_PORT;
+  return {
+    targetApp,
+    metroHost,
+    metroPort,
+    kind,
+    shouldProbeMetro: shouldProbeMetro(req, kind),
+  };
+}
+
+export function remoteConnectionChecks(req: DaemonRequest): DoctorCheck[] {
+  const evidence = remoteConnectionEvidence(req);
+  if (!evidence) return [];
+  return [
+    {
+      id: 'remote-connection',
+      status: 'info',
+      summary: 'Remote daemon/session scope is active.',
+      evidence,
+    },
+  ];
+}
+
+export function sessionChecks(
+  sessionStore: SessionStore,
+  sessionName: string,
+  session: SessionState | undefined,
+): DoctorCheck[] {
+  const sameDeviceSessions = session
+    ? sessionStore
+        .toArray()
+        .filter(
+          (candidate) =>
+            candidate.name !== session.name &&
+            candidate.device.platform === session.device.platform &&
+            candidate.device.id === session.device.id,
+        )
+        .map((candidate) => candidate.name)
+    : [];
+
+  if (!session) {
+    return [
+      {
+        id: 'session',
+        status: 'info',
+        summary: `No active session named ${sessionName}. Doctor will use the selected device.`,
+        hint: 'This is expected before a run. Use open when app foreground state matters.',
+      },
+    ];
+  }
+
+  return [
+    {
+      id: 'session',
+      status: sameDeviceSessions.length > 0 ? 'warn' : 'pass',
+      summary:
+        sameDeviceSessions.length > 0
+          ? `Other active sessions target the same device: ${sameDeviceSessions.join(', ')}`
+          : `Active session ${session.name} targets ${session.device.name}`,
+      hint:
+        sameDeviceSessions.length > 0
+          ? 'Close stale sessions before a QA run if they belong to old attempts.'
+          : undefined,
+      command:
+        sameDeviceSessions.length > 0
+          ? `agent-device close --session ${sameDeviceSessions[0]} --platform ${session.device.platform}`
+          : undefined,
+      evidence: {
+        session: session.name,
+        sameDeviceSessions,
+        sessionStateDir: sessionStore.resolveSessionDir(session.name),
+      },
+    },
+  ];
+}
+
+function shouldProbeMetro(req: DaemonRequest, kind: DoctorKind): boolean {
+  return (
+    kind !== 'auto' ||
+    typeof req.runtime?.metroPort === 'number' ||
+    typeof req.runtime?.metroHost === 'string'
+  );
+}
+
+function remoteConnectionEvidence(req: DaemonRequest): Record<string, unknown> | undefined {
+  const configured = Object.fromEntries(
+    REMOTE_CONNECTION_FLAG_KEYS.flatMap((key) =>
+      typeof req.flags?.[key] === 'string' ? [[key, '<configured>']] : [],
+    ),
+  );
+  const evidence = {
+    ...configured,
+    ...(req.flags?.sessionIsolation === 'tenant' ? { sessionIsolation: 'tenant' } : {}),
+  };
+  return Object.keys(evidence).length > 0 ? evidence : undefined;
+}
+
+function readNonEmptyString(value: unknown): string | undefined {
+  return typeof value === 'string' && value.trim().length > 0 ? value.trim() : undefined;
+}
+
+function readPositivePort(value: unknown): number | undefined {
+  return typeof value === 'number' && Number.isInteger(value) && value > 0 ? value : undefined;
+}
