+++ src/daemon/handlers/session-doctor-types.ts
@@ -1,0 +1,20 @@
+export type DoctorStatus = 'pass' | 'warn' | 'fail' | 'info';
+
+export type DoctorKind = 'auto' | 'react-native' | 'expo';
+
+export type DoctorOptions = {
+  targetApp?: string;
+  metroHost: string;
+  metroPort: number;
+  kind: DoctorKind;
+  shouldProbeMetro: boolean;
+};
+
+export type DoctorCheck = {
+  id: string;
+  status: DoctorStatus;
+  summary: string;
+  hint?: string;
+  command?: string;
+  evidence?: Record<string, unknown>;
+};
